spa - 상태관리 뼈대
